+++ src/components/takes/TakeEntryNew.jsx
@@ -6,12 +6,14 @@ import TakeEntryLayout from './TakeEntryLayout';
  * @param {object} props
  * @param {import('@/stores/document/DocumentStore').DocumentId} props.documentId
  * @param {import('@/stores/document/DocumentStore').SceneId} props.sceneId
+ * @param {import('@/stores/document/DocumentStore').BlockId} props.blockId
  * @param {import('@/stores/document/DocumentStore').ShotId} props.shotId
  * @param {'list'|'inline'} props.viewMode
  */
 export default function TakeEntryNew({
   documentId,
   sceneId,
+  blockId,
   shotId,
   viewMode,
 }) {
@@ -27,6 +29,7 @@ export default function TakeEntryNew({
       <TakeEntryDetailsNew
         documentId={documentId}
         sceneId={sceneId}
+        blockId={blockId}
         shotId={shotId}
         viewMode={viewMode}
       />